+++ ecs/archetype.go
@@ -1,7 +1,6 @@
 package ecs
 
 import (
-	"reflect"
 	"unsafe"
 )
 
@@ -11,17 +10,15 @@ type archetype struct {
 	Ids  []ID
 	// Indirection to avoid a fixed-size array of storages
 	// Increases access time by 50-100%
-	indices    [MaskTotalBits]uint8
-	entities   storage
+	references [MaskTotalBits]*storage
+	entities   genericStorage[Entity]
 	components []storage
-	toAdd      map[ID]*archetype
-	toRemove   map[ID]*archetype
+	toAdd      []*archetype
+	toRemove   []*archetype
 }
 
-var entityType = reflect.TypeOf(Entity{})
-
 // Init initializes an archetype
-func (a *archetype) Init(capacityIncrement int, components ...componentType) {
+func (a *archetype) Init(capacityIncrement int, forStorage bool, components ...componentType) {
 	var mask BitMask
 	if len(components) > 0 {
 		a.Ids = make([]ID, len(components))
@@ -37,22 +34,22 @@ func (a *archetype) Init(capacityIncrement int, components ...componentType) {
 
 		mask.Set(c.ID, true)
 		a.Ids[i] = c.ID
-		a.indices[c.ID] = uint8(i)
 		comps[i] = storage{}
-		comps[i].Init(c.Type, capacityIncrement)
+		comps[i].Init(c.Type, capacityIncrement, forStorage)
+		a.references[c.ID] = &comps[i]
 	}
 
 	a.Mask = mask
 	a.components = comps
-	a.entities = storage{}
-	a.toAdd = map[ID]*archetype{}
-	a.toRemove = map[ID]*archetype{}
-	a.entities.Init(entityType, capacityIncrement)
+	a.entities = genericStorage[Entity]{}
+	a.toAdd = make([]*archetype, MaskTotalBits)
+	a.toRemove = make([]*archetype, MaskTotalBits)
+	a.entities.Init(capacityIncrement, forStorage)
 }
 
 // GetEntity returns the entity at the given index
 func (a *archetype) GetEntity(index uint32) Entity {
-	return *(*Entity)(a.entities.Get(index))
+	return a.entities.Get(index)
 }
 
 // Get returns the component with the given ID at the given index
@@ -60,7 +57,7 @@ func (a *archetype) Get(index uint32, id ID) unsafe.Pointer {
 	if !a.Mask.Get(id) {
 		return nil
 	}
-	return a.components[a.indices[id]].Get(index)
+	return a.references[id].Get(index)
 }
 
 // GetUnsafe returns the component with the given ID at the given index,
@@ -68,15 +65,20 @@ func (a *archetype) Get(index uint32, id ID) unsafe.Pointer {
 //
 // This is used by queries, where the entity is guaranteed to be in the archetype.
 func (a *archetype) GetUnsafe(index uint32, id ID) unsafe.Pointer {
-	return a.components[a.indices[id]].Get(index)
+	return a.references[id].Get(index)
 }
 
 // Add adds an entity with zeroed components to the archetype
-func (a *archetype) Alloc(entity Entity) uint32 {
-	idx := a.entities.Add(&entity)
+func (a *archetype) Alloc(entity Entity, zero bool) uint32 {
+	idx := a.entities.Add(entity)
 	len := len(a.components)
+
 	for i := 0; i < len; i++ {
-		a.components[i].Alloc()
+		comp := &a.components[i]
+		idx := comp.Alloc()
+		if zero {
+			comp.Zero(idx)
+		}
 	}
 	return idx
 }
@@ -86,25 +88,9 @@ func (a *archetype) Add(entity Entity, components ...Component) uint32 {
 	if len(components) != len(a.Ids) {
 		panic("Invalid number of components")
 	}
-	idx := a.entities.Add(&entity)
+	idx := a.entities.Add(entity)
 	for _, c := range components {
-		a.components[a.indices[c.ID]].Add(c.Component)
-	}
-	return idx
-}
-
-// AddPointer adds an entity with components to the archetype, using pointers
-func (a *archetype) AddPointer(entity Entity, components ...componentPointer) uint32 {
-	if len(components) != len(a.Ids) {
-		panic("Invalid number of components")
-	}
-	idx := a.entities.Add(&entity)
-	for _, c := range components {
-		if c.Pointer == nil {
-			a.components[a.indices[c.ID]].Alloc()
-		} else {
-			a.components[a.indices[c.ID]].AddPointer(c.Pointer)
-		}
+		a.references[c.ID].Add(c.Component)
 	}
 	return idx
 }
@@ -141,19 +127,29 @@ func (a *archetype) Cap() uint32 {
 
 // Set overwrites a component with the data behind the given pointer
 func (a *archetype) Set(index uint32, id ID, comp interface{}) unsafe.Pointer {
-	return a.components[a.indices[id]].Set(index, comp)
+	return a.references[id].Set(index, comp)
+}
+
+// SetPointer overwrites a component with the data behind the given pointer
+func (a *archetype) SetPointer(index uint32, id ID, comp unsafe.Pointer) unsafe.Pointer {
+	return a.references[id].SetPointer(index, comp)
+}
+
+// Zero resets th memory at the given position
+func (a *archetype) Zero(index uint32, id ID) {
+	a.references[id].Zero(index)
 }
 
 // GetTransitionAdd returns the archetype resulting from adding a component
 func (a *archetype) GetTransitionAdd(id ID) (*archetype, bool) {
-	p, ok := a.toAdd[id]
-	return p, ok
+	p := a.toAdd[id]
+	return p, p != nil
 }
 
 // GetTransitionRemove returns the archetype resulting from removing a component
 func (a *archetype) GetTransitionRemove(id ID) (*archetype, bool) {
-	p, ok := a.toRemove[id]
-	return p, ok
+	p := a.toRemove[id]
+	return p, p != nil
 }
 
 // SetTransitionAdd sets the archetype resulting from adding a component