+++ ecs/world.go
@@ -37,9 +37,11 @@ func NewWorld() World {
 
 // FromConfig creates a new [World] from a [Config]
 func FromConfig(conf Config) World {
+	entities := make([]entityIndex, 1, conf.CapacityIncrement)
+	entities[0] = entityIndex{arch: nil, index: 0}
 	w := World{
 		config:     conf,
-		entities:   []entityIndex{{arch: nil, index: 0}},
+		entities:   entities,
 		entityPool: newEntityPool(conf.CapacityIncrement),
 		bitPool:    newBitPool(),
 		registry:   newComponentRegistry(),
@@ -47,7 +49,7 @@ func FromConfig(conf Config) World {
 		locks:      BitMask{},
 		listener:   nil,
 	}
-	w.createArchetype(BitMask{})
+	w.createArchetype(BitMask{}, false)
 	return w
 }
 
@@ -68,7 +70,7 @@ func (w *World) NewEntity(comps ...ID) Entity {
 		arch = w.findOrCreateArchetype(arch, comps, nil)
 	}
 
-	idx := arch.Alloc(entity)
+	idx := arch.Alloc(entity, true)
 	len := len(w.entities)
 	if int(entity.id) == len {
 		if len == cap(w.entities) {
@@ -110,7 +112,7 @@ func (w *World) NewEntityWith(comps ...Component) Entity {
 	arch := w.archetypes.Get(0)
 	arch = w.findOrCreateArchetype(arch, ids, nil)
 
-	idx := arch.Alloc(entity)
+	idx := arch.Alloc(entity, false)
 	len := len(w.entities)
 	if int(entity.id) == len {
 		if len == cap(w.entities) {
@@ -327,25 +329,20 @@ func (w *World) Exchange(entity Entity, add []ID, rem []ID) {
 	}
 
 	oldIDs := oldArch.Components()
-	keepIDs := make([]ID, 0, len(oldIDs))
-	for _, id := range oldIDs {
-		if mask.Get(id) {
-			keepIDs = append(keepIDs, id)
-		}
-	}
 
 	arch := w.findOrCreateArchetype(oldArch, add, rem)
+	newIndex := arch.Alloc(entity, false)
 
-	allComps := make([]componentPointer, 0, len(keepIDs)+len(add))
-	for _, id := range keepIDs {
-		comp := oldArch.Get(index.index, id)
-		allComps = append(allComps, componentPointer{ID: id, Pointer: comp})
+	for _, id := range oldIDs {
+		if mask.Get(id) {
+			comp := oldArch.Get(index.index, id)
+			arch.SetPointer(newIndex, id, comp)
+		}
 	}
 	for _, id := range add {
-		allComps = append(allComps, componentPointer{ID: id, Pointer: nil})
+		arch.Zero(newIndex, id)
 	}
 
-	newIndex := arch.AddPointer(entity, allComps...)
 	swapped := oldArch.Remove(index.index)
 
 	if swapped {
@@ -432,23 +429,25 @@ func (w *World) copyTo(entity Entity, id ID, comp interface{}) unsafe.Pointer {
 func (w *World) findOrCreateArchetype(start *archetype, add []ID, rem []ID) *archetype {
 	curr := start
 	mask := start.Mask
-	for _, id := range rem {
+	maxRem := len(rem) - 1
+	maxAdd := len(add) - 1
+	for i, id := range rem {
 		mask.Set(id, false)
 		if next, ok := curr.GetTransitionRemove(id); ok {
 			curr = next
 		} else {
-			next, _ := w.findOrCreateArchetypeSlow(mask)
+			next, _ := w.findOrCreateArchetypeSlow(mask, i == maxRem && maxAdd == 0)
 			next.SetTransitionAdd(id, curr)
 			curr.SetTransitionRemove(id, next)
 			curr = next
 		}
 	}
-	for _, id := range add {
+	for i, id := range add {
 		mask.Set(id, true)
 		if next, ok := curr.GetTransitionAdd(id); ok {
 			curr = next
 		} else {
-			next, _ := w.findOrCreateArchetypeSlow(mask)
+			next, _ := w.findOrCreateArchetypeSlow(mask, i == maxAdd)
 			next.SetTransitionRemove(id, curr)
 			curr.SetTransitionAdd(id, next)
 			curr = next
@@ -457,11 +456,11 @@ func (w *World) findOrCreateArchetype(start *archetype, add []ID, rem []ID) *arc
 	return curr
 }
 
-func (w *World) findOrCreateArchetypeSlow(mask BitMask) (*archetype, bool) {
+func (w *World) findOrCreateArchetypeSlow(mask BitMask, forStorage bool) (*archetype, bool) {
 	if arch, ok := w.findArchetype(mask); ok {
 		return arch, false
 	}
-	return w.createArchetype(mask), true
+	return w.createArchetype(mask, forStorage), true
 }
 
 func (w *World) findArchetype(mask BitMask) (*archetype, bool) {
@@ -475,7 +474,7 @@ func (w *World) findArchetype(mask BitMask) (*archetype, bool) {
 	return nil, false
 }
 
-func (w *World) createArchetype(mask BitMask) *archetype {
+func (w *World) createArchetype(mask BitMask, forStorage bool) *archetype {
 	count := int(mask.TotalBitsSet())
 	types := make([]componentType, count)
 
@@ -489,7 +488,7 @@ func (w *World) createArchetype(mask BitMask) *archetype {
 	}
 	w.archetypes.Add(archetype{})
 	arch := w.archetypes.Get(w.archetypes.Len() - 1)
-	arch.Init(w.config.CapacityIncrement, types...)
+	arch.Init(w.config.CapacityIncrement, forStorage, types...)
 	return arch
 }